refactor(client): consolidate duplicate Button into ui/button
Repoint the 3 importers of the standalone @/components/button
(combobox, input-group, sonner.stories) to the canonical shadcn
@/components/ui/button, then delete the stale standalone Button and
its duplicate story. Both exported Button identically and the importers
use only Button; the canonical buttonVariants covers every size/variant
they need (ghost, icon-xs, default).

Removes fallow duplicate dup:1455c094 (button.stories.tsx vs
ui/button.stories.tsx) and leaves ui/button.stories.tsx as the single
source.

Closes #443
