+++ packages/client/src/components/combobox.tsx
@@ -8,13 +8,13 @@ import * as React from "react";
 import { Combobox as ComboboxPrimitive } from "@base-ui/react";
 import { CheckIcon, ChevronDownIcon, XIcon } from "lucide-react";
 
-import { Button } from "@/components/button";
 import {
   InputGroup,
   InputGroupAddon,
   InputGroupButton,
   InputGroupInput,
 } from "@/components/input-group";
+import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 
 const Combobox = ComboboxPrimitive.Root;
